заменена функция move folder set_acl 0.60

+++ add_ad_users_str.ps1
@@ -1,5 +1,5 @@
 #PowerShell
-#Script for restore old ad_users (students) to ad ver 0.59.1 str_fork 
+#Script for restore old ad_users (students) to ad ver 0.60 str_fork 
 
 #student info
 $str_student = "152218;Фамилия Имя Отчество;Институт б1-группа-21 2022_2027(годы обучения);17061998 (дата_рождения);z_(форма_обучения)"
@@ -82,39 +82,44 @@ function restore_user {
 }
 
 
-
-
-
-
 function move_folder{
-    $ad_home_dir = $(Get-ADUser -Identity $stud_id -Server $ad_serv -Properties HomeDirectory | Select-Object -ExpandProperty HomeDirectory)
+    param ([string]$stud_id, [string]$home_dir)   
+    $ad_home_dir = $(Get-ADUser -Identity $stud_id -Server $ad_server -Properties HomeDirectory | Select-Object -ExpandProperty HomeDirectory)
     
     #if ad_home is empty
     if ([string]::IsNullOrEmpty($ad_home_dir)){
-            create_home_dir 
+            create_folder -home_dir $home_dir
     }
     #if ad_home != home_dir
+
     elseif ($ad_home_dir.ToString().trim() -ne $home_dir){
         if (Test-Path $ad_home_dir){
             #### Write-Output "$ad_home_dir status 0" debug # #create
-            Move-Item $ad_home_dir -Destination $home_dir -Force 
-            Write-host "folder $stud_id moved socessfully"
+
+            if (Test-Path $home_dir){
+                Move-Item $ad_home_dir -Destination $home_dir -Force 
+                Write-host "folder $stud_id moved socessfully"
+            }
+            else {
+                create_folder -home_dir $home_dir
+            }
         } else {
-            create_home_dir
+            create_folder -home_dir $home_dir
         }
     }
     #if ad_home == home_dir
     elseif ($ad_home_dir.ToString().trim() -eq $home_dir) {
         if (!(Test-Path $home_dir)){
-            create_home_dir  
+            create_folder -home_dir $home_dir
         }
         else {
-            Write-Output "home_dir of $stud_id is on correct way :D"
+            Write-Output "folder of $stud_id is on correct way :D"
+         
         }
     }
 
-    set_acl 
-    set_homedrive_link 
+    set_acl -stud_id $stud_id -home_dir $home_dir
+    set_homedrive_link -stud_id $stud_id -home_dir $home_dir
 }
